requirements.txt
pytest~=8.4.2
dotenv~=0.9.9
python-dotenv~=1.1.1
Faker~=37.8.0
requests~=2.32.5
pesel~=0.4.0
phone_gen==3.0.22
python-dateutil==2.9.0.post0
allure-python-commons~=2.15.0
flake8==7.3.0